pyproject.toml
[project]
name = "seo-mcp-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.2.1",
    "ipykernel>=7.1.0",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.0.0",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.2",
    "mcp>=1.6.0",
    "nest-asyncio>=1.6.0",
    "oauth2client>=4.1.3",
    "openai-agents>=0.4.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.28.0",
]
